query-service/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
alembic>=1.13,<2.0
pika>=1.3,<2.0
python-jose[cryptography]>=3.3,<4.0
pytest>=8.3,<9.0
httpx>=0.28,<1.0
black>=25.1,<26.0
ruff>=0.11,<1.0
